Add capability to change args in the Filebeat chart
No change by default, the current args are set in the
filebeat/values.yml file.

Add the capability to overwrite the args of the Filebeat
container (in the daemonset or in the deployment). It could
be useful to change the configuration path of Filebeat, if it
is mounted by a sidecar or an initcontainer (can not be mounted
over the current /usr/share/filebeat/filebeat.yml)